+++ aimiaobi-20230801/alibabacloud_aimiaobi20230801/client.py
@@ -298,15 +298,29 @@ class Client(OpenApiClient):
         OpenApiUtilClient.convert(tmp_req, request)
         if not UtilClient.is_unset(tmp_req.color_words):
             request.color_words_shrink = OpenApiUtilClient.array_to_string_with_specified_style(tmp_req.color_words, 'ColorWords', 'json')
+        if not UtilClient.is_unset(tmp_req.stickers):
+            request.stickers_shrink = OpenApiUtilClient.array_to_string_with_specified_style(tmp_req.stickers, 'Stickers', 'json')
         body = {}
+        if not UtilClient.is_unset(request.close_music):
+            body['CloseMusic'] = request.close_music
+        if not UtilClient.is_unset(request.close_subtitle):
+            body['CloseSubtitle'] = request.close_subtitle
+        if not UtilClient.is_unset(request.close_voice):
+            body['CloseVoice'] = request.close_voice
         if not UtilClient.is_unset(request.color_words_shrink):
             body['ColorWords'] = request.color_words_shrink
+        if not UtilClient.is_unset(request.custom_voice_url):
+            body['CustomVoiceUrl'] = request.custom_voice_url
+        if not UtilClient.is_unset(request.custom_voice_volume):
+            body['CustomVoiceVolume'] = request.custom_voice_volume
         if not UtilClient.is_unset(request.height):
             body['Height'] = request.height
         if not UtilClient.is_unset(request.music_url):
             body['MusicUrl'] = request.music_url
         if not UtilClient.is_unset(request.music_volume):
             body['MusicVolume'] = request.music_volume
+        if not UtilClient.is_unset(request.stickers_shrink):
+            body['Stickers'] = request.stickers_shrink
         if not UtilClient.is_unset(request.subtitle_font_size):
             body['SubtitleFontSize'] = request.subtitle_font_size
         if not UtilClient.is_unset(request.task_id):
@@ -355,15 +369,29 @@ class Client(OpenApiClient):
         OpenApiUtilClient.convert(tmp_req, request)
         if not UtilClient.is_unset(tmp_req.color_words):
             request.color_words_shrink = OpenApiUtilClient.array_to_string_with_specified_style(tmp_req.color_words, 'ColorWords', 'json')
+        if not UtilClient.is_unset(tmp_req.stickers):
+            request.stickers_shrink = OpenApiUtilClient.array_to_string_with_specified_style(tmp_req.stickers, 'Stickers', 'json')
         body = {}
+        if not UtilClient.is_unset(request.close_music):
+            body['CloseMusic'] = request.close_music
+        if not UtilClient.is_unset(request.close_subtitle):
+            body['CloseSubtitle'] = request.close_subtitle
+        if not UtilClient.is_unset(request.close_voice):
+            body['CloseVoice'] = request.close_voice
         if not UtilClient.is_unset(request.color_words_shrink):
             body['ColorWords'] = request.color_words_shrink
+        if not UtilClient.is_unset(request.custom_voice_url):
+            body['CustomVoiceUrl'] = request.custom_voice_url
+        if not UtilClient.is_unset(request.custom_voice_volume):
+            body['CustomVoiceVolume'] = request.custom_voice_volume
         if not UtilClient.is_unset(request.height):
             body['Height'] = request.height
         if not UtilClient.is_unset(request.music_url):
             body['MusicUrl'] = request.music_url
         if not UtilClient.is_unset(request.music_volume):
             body['MusicVolume'] = request.music_volume
+        if not UtilClient.is_unset(request.stickers_shrink):
+            body['Stickers'] = request.stickers_shrink
         if not UtilClient.is_unset(request.subtitle_font_size):
             body['SubtitleFontSize'] = request.subtitle_font_size
         if not UtilClient.is_unset(request.task_id):
@@ -776,15 +804,25 @@ class Client(OpenApiClient):
             request.reference_video_shrink = OpenApiUtilClient.array_to_string_with_specified_style(tmp_req.reference_video, 'ReferenceVideo', 'json')
         if not UtilClient.is_unset(tmp_req.source_videos):
             request.source_videos_shrink = OpenApiUtilClient.array_to_string_with_specified_style(tmp_req.source_videos, 'SourceVideos', 'json')
+        if not UtilClient.is_unset(tmp_req.video_roles):
+            request.video_roles_shrink = OpenApiUtilClient.array_to_string_with_specified_style(tmp_req.video_roles, 'VideoRoles', 'json')
         body = {}
         if not UtilClient.is_unset(request.anlysis_prompt):
             body['AnlysisPrompt'] = request.anlysis_prompt
+        if not UtilClient.is_unset(request.face_identity_similarity_min_score):
+            body['FaceIdentitySimilarityMinScore'] = request.face_identity_similarity_min_score
         if not UtilClient.is_unset(request.reference_video_shrink):
             body['ReferenceVideo'] = request.reference_video_shrink
+        if not UtilClient.is_unset(request.remove_subtitle):
+            body['RemoveSubtitle'] = request.remove_subtitle
         if not UtilClient.is_unset(request.source_videos_shrink):
             body['SourceVideos'] = request.source_videos_shrink
         if not UtilClient.is_unset(request.split_interval):
             body['SplitInterval'] = request.split_interval
+        if not UtilClient.is_unset(request.video_roles_shrink):
+            body['VideoRoles'] = request.video_roles_shrink
+        if not UtilClient.is_unset(request.video_shot_face_identity_count):
+            body['VideoShotFaceIdentityCount'] = request.video_shot_face_identity_count
         if not UtilClient.is_unset(request.workspace_id):
             body['WorkspaceId'] = request.workspace_id
         req = open_api_models.OpenApiRequest(
@@ -825,15 +863,25 @@ class Client(OpenApiClient):
             request.reference_video_shrink = OpenApiUtilClient.array_to_string_with_specified_style(tmp_req.reference_video, 'ReferenceVideo', 'json')
         if not UtilClient.is_unset(tmp_req.source_videos):
             request.source_videos_shrink = OpenApiUtilClient.array_to_string_with_specified_style(tmp_req.source_videos, 'SourceVideos', 'json')
+        if not UtilClient.is_unset(tmp_req.video_roles):
+            request.video_roles_shrink = OpenApiUtilClient.array_to_string_with_specified_style(tmp_req.video_roles, 'VideoRoles', 'json')
         body = {}
         if not UtilClient.is_unset(request.anlysis_prompt):
             body['AnlysisPrompt'] = request.anlysis_prompt
+        if not UtilClient.is_unset(request.face_identity_similarity_min_score):
+            body['FaceIdentitySimilarityMinScore'] = request.face_identity_similarity_min_score
         if not UtilClient.is_unset(request.reference_video_shrink):
             body['ReferenceVideo'] = request.reference_video_shrink
+        if not UtilClient.is_unset(request.remove_subtitle):
+            body['RemoveSubtitle'] = request.remove_subtitle
         if not UtilClient.is_unset(request.source_videos_shrink):
             body['SourceVideos'] = request.source_videos_shrink
         if not UtilClient.is_unset(request.split_interval):
             body['SplitInterval'] = request.split_interval
+        if not UtilClient.is_unset(request.video_roles_shrink):
+            body['VideoRoles'] = request.video_roles_shrink
+        if not UtilClient.is_unset(request.video_shot_face_identity_count):
+            body['VideoShotFaceIdentityCount'] = request.video_shot_face_identity_count
         if not UtilClient.is_unset(request.workspace_id):
             body['WorkspaceId'] = request.workspace_id
         req = open_api_models.OpenApiRequest(